requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
curl-cffi==0.5.10
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
faker==20.1.0
capsolver==1.0.0
requests==2.31.0
pydantic==2.5.0